Convert forceLightningEventCreate
