+++ packages/salesforcedx-vscode-core/src/commands/index.ts
@@ -44,7 +44,9 @@ export {
 export {
   forceLightningComponentCreate
 } from './templates/forceLightningComponentCreate';
-export { forceLightningEventCreate } from './forceLightningEventCreate';
+export {
+  forceLightningEventCreate
+} from './templates/forceLightningEventCreate';
 export { forceLightningInterfaceCreate } from './forceLightningInterfaceCreate';
 export { forceDebuggerStop } from './forceDebuggerStop';
 export { forceConfigList } from './forceConfigList';
