+++ packages/salesforcedx-vscode-core/src/commands/templates/forceLightningEventCreate.ts
@@ -1,0 +1,58 @@
+/*
+ * Copyright (c) 2019, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+import {
+  Command,
+  SfdxCommandBuilder
+} from '@salesforce/salesforcedx-utils-vscode/out/src/cli';
+import { DirFileNameSelection } from '@salesforce/salesforcedx-utils-vscode/out/src/types';
+import { nls } from '../../messages';
+import {
+  CompositeParametersGatherer,
+  LightningFilePathExistsChecker,
+  SelectFileName,
+  SelectOutputDir,
+  SfdxCommandlet,
+  SfdxWorkspaceChecker
+} from '../commands';
+import { BaseTemplateCommand } from './baseTemplateCommand';
+
+class ForceLightningEventCreateExecutor extends BaseTemplateCommand {
+  public build(data: DirFileNameSelection): Command {
+    return new SfdxCommandBuilder()
+      .withDescription(nls.localize('force_lightning_event_create_text'))
+      .withArg('force:lightning:event:create')
+      .withFlag('--eventname', data.fileName)
+      .withFlag('--outputdir', data.outputdir)
+      .withLogName('force_lightning_event_create')
+      .build();
+  }
+
+  public createSubDirectory(): boolean {
+    return true;
+  }
+
+  public getFileExtension(): string {
+    return '.evt';
+  }
+}
+
+const fileNameGatherer = new SelectFileName();
+const outputDirGatherer = new SelectOutputDir('aura', true);
+
+export async function forceLightningEventCreate() {
+  const commandlet = new SfdxCommandlet(
+    new SfdxWorkspaceChecker(),
+    new CompositeParametersGatherer<DirFileNameSelection>(
+      fileNameGatherer,
+      outputDirGatherer
+    ),
+    new ForceLightningEventCreateExecutor(),
+    new LightningFilePathExistsChecker()
+  );
+  await commandlet.run();
+}
